[names-refactor] Deleted monthly contributions reducers and actions.
